1) prompt the variable to count the memory allocated to integer(kind=8) from integer (kind=4), as WRFDA will report negative number with big domain. M      frame/module_alloc_space.h M      frame/module_domain.F
2) Turn off the boundary check in WRFDA, as 4DVar  doesn't need the second time boundary.
M      share/input_wrf.F

3) Add CPPFLAGS (especially, -traditional) to preprocessor for bufr, blas and lapack, on some machine, cpp generate wrong staff in codes.
M       var/build/da.make
M       var/external/bufr/preproc.sh
M       Makefile

4) gen_be_stage2_gsi.exe need the MPI lib on SGI ALTIX.
M       var/build/gen_be.make


git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@4462 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d

+++ frame/module_alloc_space.h
@@ -37,7 +37,7 @@
       ! false otherwise (all allocated, modulo tl above)
       LOGICAL , INTENT(IN)            :: inter_domain_in
 
-      INTEGER , INTENT(INOUT)         :: num_bytes_allocated
+      INTEGER(KIND=8) , INTENT(INOUT)         :: num_bytes_allocated
 
 
       !  Local data.
